renamed api name to make it readable
